backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
jinja2==3.1.3
fpdf2==2.7.9
markdown==3.6
python-multipart==0.0.9
requests==2.31.0
anthropic==0.25.8
